MagicKnightAttack
I hope this is correct. It took me like 3 hours trying to remember/figure out how to set up this animation correctly so if its wrong or broken for some reason let me know what is is because WOW I am out of my depth lol. Also out of pure fear I checked that this was uploaded to the CORRECT BRANCH as to not decimate the fuckin files.